+++ crates/texlab/src/server/options.rs
@@ -107,6 +107,7 @@ pub struct DiagnosticsOptions {
 pub struct InlayHintOptions {
     pub label_definitions: Option<bool>,
     pub label_references: Option<bool>,
+    pub citations: Option<bool>,
 }
 
 #[derive(Debug, Clone, Default, Serialize, Deserialize)]
@@ -254,6 +255,7 @@ impl From<Options> for Config {
 
         config.inlay_hints.label_definitions = value.inlay_hints.label_definitions.unwrap_or(true);
         config.inlay_hints.label_references = value.inlay_hints.label_references.unwrap_or(true);
+        config.inlay_hints.citations = value.inlay_hints.citations.unwrap_or(false);
 
         config.completion.matcher = match value.completion.matcher {
             CompletionMatcher::Fuzzy => base_db::MatchingAlgo::Skim,